fire/requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-SocketIO==5.3.1
Flask-CORS==4.0.0

# Security & Authentication
bcrypt==4.0.1
werkzeug==2.3.7

# Database
pymongo==4.5.0
dnspython==2.4.2

# Image Processing & OCR
Pillow==10.0.1
opencv-python-headless==4.8.1.78
pytesseract==0.3.10

# AI/ML Dependencies
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.2
numpy==1.24.3

# Environment & Configuration
python-dotenv==1.0.0

# PDF Generation & QR Codes
reportlab==4.0.4
qrcode==7.4.2
pdfkit==1.0.0

# Communication
twilio==8.10.0
requests==2.31.0

# Production Server
gunicorn==21.2.0
eventlet==0.33.3

# Data Processing
pandas==2.1.4

# File Handling
openpyxl==3.1.2

# Form Handling
WTForms==3.1.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Core Dependencies
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
six==1.16.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4